pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "photodisarm"
version = "1.0.0"
description = "A tool for managing and processing photos"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
numpy = "^1.24.0"
opencv-python = "^4.7.0"
pillow = "^9.4.0"
rawpy = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pylint = "^2.17.0"
black = "^23.3.0"

[tool.poetry.scripts]
photodisarm = "photodisarm.ui.app:main"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
